Add support for FFNx 1.8+ (#1)
* Gardening: Cleanup, not required

* Gardening: Move dependencies to NuGet

* Update: .NET 4.8

* Update; NuGet Dependencies

* Settings: Add sfx and voice to the default watch directories

* Cleanup: FFNx is going to be downloaded if required

* FFNx: FFNx.cfg -> FFNx.toml

* Gardening: Cleanup

* FFNx: Use the new FFNx.toml file as configuration file

* FFNx: Add missing Graphics options and SFX options

* Gardening: Fix Debug Log

* Gardening: Fix potential crash

* FFNx: Ensure sfx and voice folders exist

* 7th: Ensure ModFolder overrides logic works for any monitored path

Not only data

* FFNx: Replace fancy_transparency with enable_animated_textures

* FFNx: Add Update functionality

This commit will:
- Add a new section in the General Settings window
- Allow the user to choose between multiple FFNx release channels
- Allow the user to manually check for updates and install/replace the current installed version ( if newer or older than the currently installed )

* Gardening: Update NuGet packages

* FFNx: Allow 7th to configure only known options in FFNx.toml

Preserve manual edited flags and override only the ones configured via GUI

* FFNx: Ensure the latest stable version is downloaded if not present when hitting Play

* Gardening: Inform the user how to fix the error

* FFNx: Improve UX for the end user when downloading a release

This commit will switch to the Browse panel because it contains the download widget, which will contain our FFNx release download progress.
Once the download and the extraction is completed successfully, it will switch back to the Mods panel. Purely a workaround for a poor Download UX concept.

+++ TurBoLog/DebugMonitor.cs
@@ -231,7 +231,7 @@ namespace TurBoLog {
 			try {
 				// Everything after the first DWORD is our debugging text
 				IntPtr pString = new IntPtr(
-					m_SharedMem.ToInt32() + Marshal.SizeOf(typeof(int))
+					m_SharedMem.ToInt64() + Marshal.SizeOf(typeof(int))
 				);
 
 				while (true) {
